# frontend/api/默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/attendance/session/start": {
      "post": {
        "summary": "老师发起一次签到场次",
        "deprecated": false,
        "description": "POST /api/attendance/session/start",
        "tags": [
          "考勤 / 签到相关接口"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceSessionStartRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAttendanceSessionStartResponse"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "sessionId": 0,
                    "token": "",
                    "allowStartTime": "",
                    "allowEndTime": "",
                    "tokenExpireTime": ""
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/attendance/session/{sessionId}/close": {
      "post": {
        "summary": "老师关闭本次签到场次，并生成缺勤记录",
        "deprecated": false,
        "description": "POST /api/attendance/session/{sessionId}/close?teacherId=1001",
        "tags": [
          "考勤 / 签到相关接口"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "teacherId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/attendance/checkin": {
      "post": {
        "summary": "学生扫码 / 输入口令签到",
        "deprecated": false,
        "description": "POST /api/attendance/checkin",
        "tags": [
          "考勤 / 签到相关接口"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCheckinRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/attendance/teacher/sessions": {
      "get": {
        "summary": "老师查看某段时间内自己的签到场次列表",
        "deprecated": false,
        "description": "GET /api/attendance/teacher/sessions?teacherId=1001&startDate=2024-10-01&endDate=2024-10-31",
        "tags": [
          "考勤 / 签到相关接口"
        ],
        "parameters": [
          {
            "name": "teacherId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListAttendanceSessionSummaryDTO"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "sessionId": 0,
                      "offeringId": 0,
                      "courseName": "",
                      "className": "",
                      "courseDate": "",
                      "sectionStart": 0,
                      "sectionEnd": 0,
                      "status": "",
                      "allowStartTime": "",
                      "allowEndTime": "",
                      "shouldAttendCount": 0,
                      "checkedInCount": 0,
                      "leaveCount": 0
                    }
                  ]
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/attendance/session/{sessionId}/detail": {
      "get": {
        "summary": "老师查看某一次签到场次详情",
        "deprecated": false,
        "description": "GET /api/attendance/session/{sessionId}/detail?teacherId=1001",
        "tags": [
          "考勤 / 签到相关接口"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "teacherId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAttendanceSessionDetailDTO"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "sessionId": 0,
                    "offeringId": 0,
                    "courseName": "",
                    "className": "",
                    "courseDate": "",
                    "sectionStart": 0,
                    "sectionEnd": 0,
                    "status": "",
                    "allowStartTime": "",
                    "allowEndTime": "",
                    "shouldAttendCount": 0,
                    "checkedInCount": 0,
                    "leaveCount": 0,
                    "absenceCount": 0,
                    "students": [
                      {
                        "studentId": 0,
                        "studentNo": "",
                        "studentName": "",
                        "status": "",
                        "checkinTime": "",
                        "leaveId": 0
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttendanceSessionStartResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "token": {
            "type": "string",
            "description": "本次签到的token，前端可以编码到二维码中"
          },
          "allowStartTime": {
            "type": "string",
            "description": ""
          },
          "allowEndTime": {
            "type": "string",
            "description": ""
          },
          "tokenExpireTime": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "ResultAttendanceSessionStartResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/AttendanceSessionStartResponse",
            "description": ""
          }
        }
      },
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "description": "",
            "type": "null"
          }
        }
      },
      "AttendanceSessionStartRequest": {
        "type": "object",
        "properties": {
          "teacherId": {
            "type": "integer",
            "description": "任课老师ID（Staff.staffId）",
            "format": "int64"
          },
          "offeringId": {
            "type": "integer",
            "description": "开课ID（Offering.offeringId）",
            "format": "int64"
          },
          "courseDate": {
            "type": "string",
            "description": "上课日期"
          },
          "sectionStart": {
            "type": "integer",
            "description": "节次范围"
          },
          "sectionEnd": {
            "type": "integer",
            "description": ""
          },
          "durationMinutes": {
            "type": "integer",
            "description": "签到允许时长（分钟），例如 10 分钟"
          }
        }
      },
...
      "StudentCheckinRequest": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "description": "学生ID（Student.studentId）",
            "format": "int64"
          },
          "token": {
            "type": "string",
            "description": "签到token（来自二维码）"
          }
        }
      },
...
      "AttendanceSessionSummaryDTO": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "offeringId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "courseName": {
            "type": "string",
            "description": ""
          },
          "className": {
            "type": "string",
            "description": ""
          },
          "courseDate": {
            "type": "string",
            "description": ""
          },
          "sectionStart": {
            "type": "integer",
            "description": ""
          },
          "sectionEnd": {
            "type": "integer",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": "OPEN / CLOSED"
          },
          "allowStartTime": {
            "type": "string",
            "description": ""
          },
          "allowEndTime": {
            "type": "string",
            "description": ""
          },
          "shouldAttendCount": {
            "type": "integer",
            "description": ""
          },
          "checkedInCount": {
            "type": "integer",
            "description": ""
          },
          "leaveCount": {
            "type": "integer",
            "description": ""
          }
        }
      },
...
      "ResultListAttendanceSessionSummaryDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttendanceSessionSummaryDTO",
              "description": "老师查看某段时间内的签到场次列表"
            },
            "description": ""
          }
        }
      },
      "StudentStatusItem": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "studentNo": {
            "type": "string",
            "description": ""
          },
          "studentName": {
            "type": "string",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": "PRESENT / ABSENT / LEAVE"
          },
          "checkinTime": {
            "type": "string",
            "description": ""
          },
          "leaveId": {
            "type": "integer",
            "description": "如果是请假，关联的 leaveId（可选，看你是否需要）",
            "format": "int64"
          }
        }
      },
...
      "AttendanceSessionDetailDTO": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "offeringId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "courseName": {
            "type": "string",
            "description": ""
          },
          "className": {
            "type": "string",
            "description": ""
          },
          "courseDate": {
            "type": "string",
            "description": ""
          },
          "sectionStart": {
            "type": "integer",
            "description": ""
          },
          "sectionEnd": {
            "type": "integer",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": "OPEN / CLOSED"
          },
          "allowStartTime": {
            "type": "string",
            "description": ""
          },
          "allowEndTime": {
            "type": "string",
            "description": ""
          },
          "shouldAttendCount": {
            "type": "integer",
            "description": ""
          },
          "checkedInCount": {
            "type": "integer",
            "description": ""
          },
          "leaveCount": {
            "type": "integer",
            "description": ""
          },
          "absenceCount": {
            "type": "integer",
            "description": ""
          },
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentStatusItem",
              "description": "com.example.leavesystem.dto.AttendanceSessionDetailDTO.StudentStatusItem"
            },
            "description": ""
          }
        }
      },
...
      "ResultAttendanceSessionDetailDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/AttendanceSessionDetailDTO",
            "description": ""
          }
        }
      },
...
    },
...
    "securitySchemes": {}
  },
  "servers": [
    {
      "url": "http://dev-cn.your-api-server.com",
      "description": "开发环境"
    }
  ],
  "security": []
}